+++ Realm/RLMObjectSchema.mm
@@ -68,10 +68,8 @@
         BOOL ignored = [[objectClass ignoredProperties] containsObject:propertyName];
         
         if (!ignored) { // Don't process ignored properties
-            RLMProperty *prop = [RLMProperty propertyForObjectProperty:props[i]
-                                                            attributes:[objectClass attributesForProperty:propertyName]
-                                                                column:propArray.count];
-            
+            RLMPropertyAttributes attr = [objectClass attributesForProperty:propertyName];
+            RLMProperty *prop = [RLMProperty propertyForObjectProperty:props[i] attributes:attr];
             if (prop) {
                 [propArray addObject:prop];
             }