+++ Realm/RLMRealm.h
@@ -93,6 +93,22 @@
  */
 @property (nonatomic, readonly) RLMSchema *schema;
 
+/**---------------------------------------------------------------------------------------
+ *  @name Default Realm Path
+ * ---------------------------------------------------------------------------------------
+ */
+/**
+ Returns the location of the default Realm as a string.
+
+ `~/Documents/default.realm` on OSX.
+
+ `default.realm` in your application's documents directory on iOS.
+
+ @return Location of the default Realm.
+
+ @see defaultRealm
+ */
++ (NSString *)defaultRealmPath;
 
 #pragma mark - Notifications